Fix usage item showing even without trigger
Returns nothing if query isn't triggered, returning an Item was causing
it to always be displayed, even when there was no query at all.

Also waits to return the usage Item until the config has been "verified"
so that we don't need to start typing something to translate just for it
to tell us we're missing the project ID in our config.