Handle exceptions when trying to load API client
If we instantiate the client with an invalid or missing service key
path, it bails out with a DefaultCredentialsError exception.

Excepting that particular thing doesn't work and I'm too lazy to hunt it
down in the API source, so this just catches all exceptions and adds a
check to display a friendly-ish message if the client wasn't loaded.