Refactor translation handling
Moves the bulk of the handleQuery method into its own translate function
and simplifies the exception handling so that it only spits out one line
in the console instead of a warning AND the printed exception.

The actual conversion of an API response into an Item is also moved into
a self-contained function to keep it all together in one place.

tl;dr: Made shit more readable and maintainable.