Handle invalid language codes
Replaces `lang` with `target` in the for loop so that we don't override
the Lang class, and adds a new `has` method which we use to check
whether the key exists before attempting to translate to a language.

Previously the API would catch it and trigger a GoogleAPICallError so it
wasn't that big of an issue, but this should speed things up by removing
the need for an API request for any invalid target languages.