Add ability to set source language
Reworks handling of `to:foo` to loop through the first two parts of the
string, checking each part for either `from` or `to` arguments so that
they can be passed in any order.

If nothing is specified after the `:` we skip to avoid setting an empty
source/target language, and if there's no colon at all we simply break
out of the loop. If args are passed, they MUST be before anything else.

Updates the API call to have params passed in via a dict to allow adding
the source language only if it's set, because "auto" isn't really valid.

The detected_language_code key also doesn't seem to be set if source is
set manually, so we set the subtext based on whether source is "auto".