Hide translations that match detected language
This doesn't have an affect if there is only one target language defined
in config. It is most useful with two targets defined, such as `en,fr`,
where it will enable to/from either language without setting to or from
args in the actual query to override anything.

Before this commit, it would still be possible but `tr Hello` would
yield two results: one en-fr with "Bonjour", but also en-en with
"Hello", which you could argue is slightly redundant.

Does not apply if source is manually set to a specific language, because
otherwise there would be no results at all. Instead, the API will throw
an error which will be displayed in the single returned item.